api/requirements.txt
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
essential-generators==1.0
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
mypy-extensions==0.4.3
names==0.3.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
python-decouple==3.5
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
random-username==1.0.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
tomli==1.2.3
typing-extensions==4.0.1
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1